+++ src/main/java/gtruty/sanescobar/entities/PlaceOfStart/BuildingsEntityNorth.java
@@ -34,9 +34,9 @@ public class BuildingsEntityNorth {
     @Column(name = "info")
     private String info;
     @Column(name = "number_of_animals")
-    private Long numberOfAnimal;
+    private Integer numberOfAnimal;
     @Column(name = "place_for_goods")
-    private Long placeForGoods;
+    private Integer placeForGoods;
     @Column(name = "special_for")
     private String specialFor;
     @Column (name = "amaount")
@@ -53,7 +53,7 @@ public class BuildingsEntityNorth {
         this.name = name;
     }
 
-    public BuildingsEntityNorth(String name, String products1, String products2, BigDecimal moneyProfit1, BigDecimal moneyProfit2, BigDecimal costOfBuilding, String requirements1, String requirements2, String suistenance, Double valueOfSuistenance, String info, Long numberOfAnimal, Long placeForGoods, String specialFor, Integer amount, Integer type) {
+    public BuildingsEntityNorth(String name, String products1, String products2, BigDecimal moneyProfit1, BigDecimal moneyProfit2, BigDecimal costOfBuilding, String requirements1, String requirements2, String suistenance, Double valueOfSuistenance, String info, Integer numberOfAnimal, Integer placeForGoods, String specialFor, Integer amount, Integer type) {
         this.name = name;
         this.products1 = products1;
         this.products2 = products2;
@@ -72,7 +72,7 @@ public class BuildingsEntityNorth {
         this.type = type;
     }
 
-    public BuildingsEntityNorth(String name, String products1, String products2, BigDecimal moneyProfit1, BigDecimal moneyProfit2, BigDecimal costOfBuilding, String requirements1, String requirements2, String suistenance, Double valueOfSuistenance, String info, Long numberOfAnimal, Long placeForGoods, String specialFor, Integer amount, Integer type, String animalInBuilding) {
+    public BuildingsEntityNorth(String name, String products1, String products2, BigDecimal moneyProfit1, BigDecimal moneyProfit2, BigDecimal costOfBuilding, String requirements1, String requirements2, String suistenance, Double valueOfSuistenance, String info, Integer numberOfAnimal, Integer placeForGoods, String specialFor, Integer amount, Integer type, String animalInBuilding) {
         this.name = name;
         this.products1 = products1;
         this.products2 = products2;
@@ -188,19 +188,19 @@ public class BuildingsEntityNorth {
         this.info = info;
     }
 
-    public Long getNumberOfAnimal() {
+    public Integer getNumberOfAnimal() {
         return numberOfAnimal;
     }
 
-    public void setNumberOfAnimal(Long numberOfAnimal) {
+    public void setNumberOfAnimal(Integer numberOfAnimal) {
         this.numberOfAnimal = numberOfAnimal;
     }
 
-    public Long getPlaceForGoods() {
+    public Integer getPlaceForGoods() {
         return placeForGoods;
     }
 
-    public void setPlaceForGoods(Long placeForGoods) {
+    public void setPlaceForGoods(Integer placeForGoods) {
         this.placeForGoods = placeForGoods;
     }
