+++ src/main/java/gtruty/sanescobar/entities/PlaceOfStart/BuildingsEntitySouth.java
@@ -35,9 +35,9 @@ public class BuildingsEntitySouth {
     @Column(name = "info")
     private String info;
     @Column(name = "number_of_animals")
-    private Long numberOfAnimal;
+    private Integer numberOfAnimal;
     @Column(name = "place_for_goods")
-    private Long placeForGoods;
+    private Integer placeForGoods;
     @Column(name = "special_for")
     private String specialFor;
     @Column (name = "amaount")
@@ -54,7 +54,7 @@ public class BuildingsEntitySouth {
         this.name = name;
     }
 
-    public BuildingsEntitySouth(String name, String products1, String products2, BigDecimal moneyProfit1, BigDecimal moneyProfit2, BigDecimal costOfBuilding, String requirements1, String requirements2, String suistenance, Double valueOfSuistenance, String info, Long numberOfAnimal, Long placeForGoods, String specialFor, Integer amount, Integer type) {
+    public BuildingsEntitySouth(String name, String products1, String products2, BigDecimal moneyProfit1, BigDecimal moneyProfit2, BigDecimal costOfBuilding, String requirements1, String requirements2, String suistenance, Double valueOfSuistenance, String info, Integer numberOfAnimal, Integer placeForGoods, String specialFor, Integer amount, Integer type) {
         this.name = name;
         this.products1 = products1;
         this.products2 = products2;
@@ -73,7 +73,7 @@ public class BuildingsEntitySouth {
         this.type = type;
     }
 
-    public BuildingsEntitySouth(String name, String products1, String products2, BigDecimal moneyProfit1, BigDecimal moneyProfit2, BigDecimal costOfBuilding, String requirements1, String requirements2, String suistenance, Double valueOfSuistenance, String info, Long numberOfAnimal, Long placeForGoods, String specialFor, Integer amount, Integer type, String animalInBuilding) {
+    public BuildingsEntitySouth(String name, String products1, String products2, BigDecimal moneyProfit1, BigDecimal moneyProfit2, BigDecimal costOfBuilding, String requirements1, String requirements2, String suistenance, Double valueOfSuistenance, String info, Integer numberOfAnimal, Integer placeForGoods, String specialFor, Integer amount, Integer type, String animalInBuilding) {
         this.name = name;
         this.products1 = products1;
         this.products2 = products2;
@@ -189,19 +189,19 @@ public class BuildingsEntitySouth {
         this.info = info;
     }
 
-    public Long getNumberOfAnimal() {
+    public Integer getNumberOfAnimal() {
         return numberOfAnimal;
     }
 
-    public void setNumberOfAnimal(Long numberOfAnimal) {
+    public void setNumberOfAnimal(Integer numberOfAnimal) {
         this.numberOfAnimal = numberOfAnimal;
     }
 
-    public Long getPlaceForGoods() {
+    public Integer getPlaceForGoods() {
         return placeForGoods;
     }
 
-    public void setPlaceForGoods(Long placeForGoods) {
+    public void setPlaceForGoods(Integer placeForGoods) {
         this.placeForGoods = placeForGoods;
     }
